# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HilmiAdinko API",
    "version": "1.0.0",
    "description": "REST API for the HilmiAdinko portfolio, contact, kategori layanan, and testimoni management system.\n\nDibuat oleh **Ahmad Hilmi Fahrezi**.",
    "contact": {
      "name": "Ahmad Hilmi Fahrezi"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/testimoni": {
      "get": {
        "tags": [
          "Testimoni"
        ],
        "summary": "Get all testimoni (legacy)",
        "operationId": "getTestimoni",
        "responses": {
          "200": {
            "description": "Testimoni retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Testimoni data has been retrieved"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Testimoni"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/testimoniRoute": {
      "get": {
        "tags": [
          "TestimoniCRUD"
        ],
        "summary": "Get all testimoni (full CRUD route)",
        "operationId": "getAllTestimoni",
        "responses": {
          "200": {
            "description": "List of testimoni",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Get all testimoni success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Testimoni"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": [
          "TestimoniCRUD"
        ],
        "summary": "Create a new testimoni",
        "operationId": "createNewTestimoni",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestimoniInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Testimoni created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Create new testimoni success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/TestimoniInput"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/testimoniRoute/{id}": {
      "put": {
        "tags": [
          "TestimoniCRUD"
        ],
        "summary": "Update a testimoni",
        "operationId": "updateTestimoni",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestimoniInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Testimoni updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Update testimoni success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Testimoni"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "TestimoniCRUD"
        ],
        "summary": "Delete a testimoni",
        "operationId": "deleteTestimoni",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Testimoni deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Delete testimoni success"
                    },
                    "data": {
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Testimoni": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "nama_klien": {
            "type": "string",
            "example": "Budi Santoso"
          },
          "waktu": {
            "type": "string",
            "format": "date",
            "example": "2024-01-15"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "example": 5
          },
          "deskripsi": {
            "type": "string",
            "example": "Pelayanan sangat memuaskan dan profesional."
          }
        }
      },
      "TestimoniInput": {
        "type": "object",
        "required": [
          "nama_klien",
          "waktu",
          "rating",
          "deskripsi"
        ],
        "properties": {
          "nama_klien": {
            "type": "string",
            "example": "Budi Santoso"
          },
          "waktu": {
            "type": "string",
            "format": "date",
            "example": "2024-01-15"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "example": 5
          },
          "deskripsi": {
            "type": "string",
            "example": "Pelayanan sangat memuaskan dan profesional."
          }
        }
      },
...
    },
    "responses": {
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Server Error"
                },
                "serverMessage": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  }
}